feat(boilerplate): update align typescript support
