+++ package.json
@@ -7,7 +7,10 @@
     "ios": "react-native run-ios",
     "start": "react-native start",
     "test": "jest",
-    "lint": "eslint ."
+    "lint": "eslint .",
+    "clear-cache": "react-native clear-cache",
+    "clear-logs": "react-native log-android --clear",
+    "reinstall-node-modules": "rm -rf node_modules && yarn install"
   },
   "dependencies": {
     "@react-native-async-storage/async-storage": "^1.17.0",
