+++ src/Screens/SplashScreen/RenderSplashScreen.type.ts
@@ -1,3 +1,6 @@
+export interface IndexProps {
+    navigation: any;
+}
 export interface Props {
     navigation: any;
     token: string;
