+++ src/Services/Redux/Reducers/AuthReducer.js
@@ -1,3 +1,4 @@
+/* eslint-disable import/extensions */
 /* eslint-disable no-underscore-dangle */
 import Action from '../ConstantReducer';
